Add DeckInitializer service for initializing user decks and cards
