+++ src/main/java/com/wordonline/matching/auth/service/UserService.java
@@ -10,6 +10,7 @@ import com.wordonline.matching.auth.domain.UserStatus;
 import com.wordonline.matching.auth.dto.UserDetailResponseDto;
 import com.wordonline.matching.auth.dto.UserResponseDto;
 import com.wordonline.matching.auth.repository.UserRepository;
+import com.wordonline.matching.deck.service.DeckInitializer;
 import com.wordonline.matching.deck.service.DeckService;
 import com.wordonline.matching.matching.client.AccountClient;
 import com.wordonline.matching.service.LocalizationService;
@@ -27,6 +28,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final DeckService deckService;
+    private final DeckInitializer deckInitializer;
     private final AccountClient accountClient;
     private final LocalizationService localizationService;
 
@@ -43,7 +45,7 @@ public class UserService {
                         );
                     })
                 ).flatMap(saveUser ->
-                    deckService.initializeCard(saveUser.getId())
+                        deckInitializer.initializeCard(saveUser.getId())
                             .map(deckId -> Tuples.of(saveUser, deckId))
                 ).flatMap(tuple -> {
                     tuple.getT1().setSelectedDeckId(tuple.getT2());
